fix(e2e): remove duplicate functions rel. to setup/teardown of node-red
